eel2: fix warnings -- from 552a7902

+++ WDL/eel2/nseel-compiler.c
@@ -831,7 +831,7 @@ static void *__newBlock(llBlock **start, int size, int wantMprotect)
       static int pagesize = 0;
       if (!pagesize)
       {
-        pagesize=sysconf(_SC_PAGESIZE);
+        pagesize=(int)sysconf(_SC_PAGESIZE);
         if (!pagesize) pagesize=4096;
       }
       uintptr_t offs,eoffs;
@@ -5601,7 +5601,7 @@ opcodeRec *nseel_translate(compileContext *ctx, const char *tmp, size_t tmplen) 
     if (tmp[1] == '~')
     {
       char *p=(char*)tmp+2;
-      unsigned int v=strtoul(tmp+2,&p,10);
+      unsigned int v=(unsigned int) strtoul(tmp+2,&p,10);
       if (v>53) v=53;
       return nseel_createCompiledValue(ctx,(EEL_F)((((WDL_INT64)1) << v) - 1));
     }
